requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
pycryptodome==3.18.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0